nbt: add & fix nesting depth limit checks
The previous nesting depth check implentation was faulty in that it was checking how many elements were being added & not how many times it had 'nested' (recursed) between either nbt_list() or nbt_compound(). I do not know why the value of 512 is the maximum depth allowed for nesting, just that the wiki specifies so.