nbt: fix LArrays not parsing a name
consume() has had a slight error the entire time in that the cases it matches to do a nbt_string() parse sequence is based on the range operator. However, I incorrectly put the match sequence as '0..12' which translate to Range { start: 0, end: 12 } so 12 is our _END_ valuei.e it won't satify with the value 0xc (12) since it doesn't lie between the start & end.
TLDR; off by one in the name parsing condition for LArray (0xC - 12 opcode).