rave: begin CLI implementation
This will remove all of the previous placeholder that built up while testing NBT and then Region parsing. We want there to be an agnostic layer underneath that can be used as a crate, cli, or even in the extremely unlikely case externally linked. For now however, we will simply parse the arguments, command and subcommand if applicable and do nothing with them.